+++ lib/src/app/app_command_runner.dart
@@ -5,6 +5,7 @@ import 'package:dvmx/src/app/commands/dart_command.dart';
 import 'package:dvmx/src/app/commands/install_command.dart';
 import 'package:dvmx/src/app/commands/list_command.dart';
 import 'package:dvmx/src/app/commands/releases_command.dart';
+import 'package:dvmx/src/app/commands/uninstall_command.dart';
 import 'package:dvmx/src/app/commands/use_command.dart';
 import 'package:dvmx/src/app/gen/cli_info.g.dart';
 import 'package:dvmx/src/app/models/exit_status.dart';
@@ -33,6 +34,7 @@ final class AppCommandRunner extends CommandRunner<ExitStatus> {
     addCommand(UseCommand());
     addCommand(DartCommand());
     addCommand(ListCommand());
+    addCommand(UninstallCommand());
   }
 
   ConsoleService get _consoleService =>