+++ lib/src/app/command_services/uninstall_command_service.g.dart
@@ -1,0 +1,35 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+// ignore_for_file: type=lint, duplicate_ignore
+
+part of 'uninstall_command_service.dart';
+
+// **************************************************************************
+// RiverpodGenerator
+// **************************************************************************
+
+String _$uninstallCommandServiceHash() =>
+    r'173fb31f0a549946daf411805f76b686a23278bb';
+
+/// See also [uninstallCommandService].
+@ProviderFor(uninstallCommandService)
+final uninstallCommandServiceProvider =
+    AutoDisposeProvider<UninstallCommandService>.internal(
+  uninstallCommandService,
+  name: r'uninstallCommandServiceProvider',
+  debugGetCreateSourceHash: const bool.fromEnvironment('dart.vm.product')
+      ? null
+      : _$uninstallCommandServiceHash,
+  dependencies: <ProviderOrFamily>[sdkServiceProvider, consoleServiceProvider],
+  allTransitiveDependencies: <ProviderOrFamily>{
+    sdkServiceProvider,
+    ...?sdkServiceProvider.allTransitiveDependencies,
+    consoleServiceProvider,
+    ...?consoleServiceProvider.allTransitiveDependencies
+  },
+);
+
+typedef UninstallCommandServiceRef
+    = AutoDisposeProviderRef<UninstallCommandService>;
+// ignore_for_file: type=lint
+// ignore_for_file: subtype_of_sealed_class, invalid_use_of_internal_member, invalid_use_of_visible_for_testing_member