+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Create a command for showing installed Dart SDK versions.
   - `dvm list`
+- Create a command for uninstalling Dart SDK versions.
+  - `dvm uninstall <version>`
 
 ## 0.0.2